test(annex-ab): split SC Result modules under cap
